공동 레이아웃 추가, swiper 설치
