+++ src/components/Banner.tsx
@@ -1,0 +1,20 @@
+import { Swiper, SwiperSlide } from "swiper/react";
+import "swiper/css";
+
+const Banner = () => {
+  return (
+    <Swiper
+      spaceBetween={50}
+      slidesPerView={1}
+      onSlideChange={() => console.log("slide change")}
+      onSwiper={(swiper) => console.log(swiper)}
+    >
+      <SwiperSlide>Slide 1</SwiperSlide>
+      <SwiperSlide>Slide 2</SwiperSlide>
+      <SwiperSlide>Slide 3</SwiperSlide>
+      <SwiperSlide>Slide 4</SwiperSlide>
+    </Swiper>
+  );
+};
+
+export default Banner;
